pyproject.toml
[project]
name = "giftool2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask",
    "opencv-python",
    "pillow>=11.2.1",
]
